Refactored away a lot of duplicated code / logic
